fix: rasterize svg
